% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dull-utils.R
\name{adding-routes}
\alias{adding-routes}
\alias{get}
\alias{method}
\alias{post}
\alias{put}
\title{Add routes}
\usage{
method(.app, method, uri, callback)

get(.app, uri, callback)

post(.app, uri, callback)

put(.app, uri, callback)
}
\arguments{
\item{.app}{A \code{dull_app} R6 class object}

\item{method}{An HTTP method, see section \bold{HTTP} Methods for details}

\item{uri}{The route URI, treated as a regular expression, see section URIs}

\item{callback}{A function of two arguments \code{request} and \code{response}, see section Callback functions for details}
}
\description{
Add a route to a dull application object. A route is comprised of an HTTP
method, a URI, and a callback function.
}
\section{HTTP Methods}{

TODO
}

\section{URIs}{

TODO
}

