% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.R
\name{all.equal.route}
\alias{all.equal.route}
\title{Test if Two Routes are (Nearly) Equal}
\usage{
\method{all}{equal.route}(target, current, ...)
}
\arguments{
\item{target, current}{Two routes to compare.}

\item{identical}{If FALSE, test for equivalency.}
}
\value{
TRUE if both routes have the same path and method. The special method
\code{"ALL"} is considered equal to any number of other methods.
}
\description{
This implementation of \code{all.equal} allows comparison of two routes.
}
\examples{
rte80 <- route(
  'GET',
  '^/route/to/ohio$',
  function(req) {
    response()  
  }
)

rte90 <- route(
  'all',
  '^/route/to/ohio$',
  function(req) {
   res <- response()
   status(res) <- 300
   res
  }
)

all.equal(rte80, rte90)
}

