% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application.R
\name{app}
\alias{app}
\title{Create application}
\usage{
app(...)
}
\arguments{
\item{\ldots}{Route objects}
}
\description{
An application in prairie is simply defined as \emph{a list of 
\link{routes}}. The application handles dispatching HTTP requests to routes 
and sending back HTTP responses. Route handlers are pure functions and 
therefore for a given HTTP request an application will always return the same
HTTP response.
}
\details{

}

