% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response-utils.R
\name{as.list.response}
\alias{as.list.response}
\title{Coerce Response to List}
\usage{
\method{as.list}{response}(x, ...)
}
\arguments{
\item{x}{A response object.}

\item{\dots}{Ignored.}
}
\description{
Used to format a response object in the acceptable httpuv format.
}
\keyword{internal}
