% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{as.request}
\alias{as.request}
\alias{as.request.environment}
\title{Coerce Requests}
\usage{
as.request(x)

\method{as.request}{environment}(envir)
}
\description{
Coerce Requests
}
\examples{
e <- new.env(parent = baseenv())

e$REQUEST_METHOD <- 'GET'
e$PATH_INFO <- '/file/download'

as.request(e)
}
\keyword{internal}

