% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{as.response}
\alias{as.response}
\alias{as.response.response}
\alias{as.response.condition}
\alias{as.response.connection}
\alias{as.response.shiny.tag}
\alias{as.response.data.frame}
\alias{as.response.matrix}
\alias{is.response}
\title{Coerce Objects to Responses}
\usage{
as.response(x, ...)

\method{as.response}{response}(x, ...)

\method{as.response}{condition}(x, ...)

\method{as.response}{connection}(x, content_type = "text/plain",
  collapse = "\\n", ...)

\method{as.response}{shiny.tag}(x, status = 200)

\method{as.response}{data.frame}(x, format = "json", ...)

\method{as.response}{matrix}(x, ...)

is.response(x)
}
\arguments{
\item{x}{Any \R object.}

\item{collapse}{A character string specifying how to collapse the lines read
from connection \code{x}.}

\item{format}{A character string, determining the form of the HTTP response.
Can be one of \code{"csv"}, \code{"html"}, \code{"json"}, or \code{"text"}.}

\item{\ldots}{Additional arguments passed on to methods.}
}
\description{
The \code{as.response} functions simplify serving up R objects as server
responses.
}
\details{
\code{as.response.character} expects \code{x} is a character string
specifying a file name. The default directory for the file is "views", but a
different path may be specified with the \code{directory} argument. If the
file exists the contents are read and set as the response body. The response
Content-Type is guessed from the file extension using
\code{\link[mime]{guess_type}}.

\code{as.response.data.frame} coerces and serves up a data frame.
Several response types are possible.
\itemize{
  \item CSV (text/csv) - Same format as \link{write.csv}, without row names.
  \item HTML (text/html) - Formatted by \link{simpleHtmlTable}.
    Additional arguments may be passed to the formatter using \code{\ldots}.
  \item JSON (application/json) - The
    data frame is coerced using the \code{\link{as.json}} function and additional
    arguments may be passed to \code{as.json} using \code{\ldots}.
  \item Text (text/plain) - Simple text rendering.
}
}
\examples{

is.response(logical(1))
is.response(response())
is.response(3030)
}
