% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route-class.R
\name{as.route}
\alias{as.route}
\alias{as.route.character}
\alias{as.route.default}
\alias{as.route.list}
\alias{as.route.route}
\title{Converting objects to routes}
\usage{
as.route(obj, ...)

\method{as.route}{route}(rte)

\method{as.route}{character}(file, path = "views")

\method{as.route}{list}(x)

\method{as.route}{default}(other)
}
\arguments{
\item{rte}{An object of class \code{rte}}

\item{file}{A file which defines a route}

\item{path}{Folder path to route file}

\item{x}{A \code{list} with items \code{method}, \code{path}, and
\code{handler}}

\item{other}{An object of an unsupported class}
}
\description{
Routes may be defined with the
}

