% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{body}
\alias{body}
\alias{body.request}
\alias{body.response}
\alias{body<-.response}
\title{HTTP Message Body}
\usage{
\method{body}{request}(req)

\method{body}{response}(res)

\method{body}{response}(res) <- value
}
\arguments{
\item{req}{A \code{request} object.}

\item{res}{A \code{response} object.}

\item{value}{Character vector to set as message body of \code{res}.}
}
\value{
A character string if the request contains a message body, otherwise 
\code{NULL}.
}
\description{
Get the message body of a \code{request} or \code{response} object. The body
of \code{response} objects may be set.
}
\examples{
route(
  'post',
  '^transmogrify/json$',
  function() {
    req <- request()
    res <- response()
    
    # get body of request
    raw_body <- body(req)
    data <- jsonlite::fromJSON(raw_body)
    
    mogrified <- lapply(
      data, 
      function(col) paste0(as.character(col), '!!!')
    )
    
    status(res) <- 200
    
    # set body of response as data.frame
    # coerced to JSON string with as.JSON
    body(res) <- mogrified
    
    res
  }
)
}
\seealso{
Additional \code{request} functions: \code{\link{body}}, \code{\link{host}}, 
\code{\link{type}}, \code{\link{url}}

For more information about HTTP requests please refer to the 
\href{http://www.w3.org/Protocols/rfc2616/rfc2616-sec5.html}{Request}
section of www.w3.org.
}

