% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dull-app.R
\name{dull}
\alias{dull}
\title{Create a new dull application object}
\usage{
dull()
}
\value{
A \code{dull_app} R6 class object.
}
\description{
Create a dull application object. The dull application object
encapsulates route URIs, callback functions, and [ADD MORE] that together form a web application.
}
\details{
See \link{dull_app} for full details of the underlying R6 object of the dull
application.
}

