% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dull-class.R
\docType{class}
\name{dull_app}
\alias{dull_app}
\title{Dull application class object}
\format{An R6 class object.}
\usage{
dull_app
}
\description{
An R6 class which is the backbone of the dull package.
}
\keyword{internal}

