% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request-header.R
\name{host}
\alias{host}
\title{HTTP Request Host}
\usage{
host(req)
}
\arguments{
\item{req}{A \code{request} object}
}
\description{
The host portion of the Host request header field.
}
\examples{
route(
  'get',
  '^.*',
  function() {
    req <- request()
    res <- response()
    
    status(res) <- 200
    body(res) <- paste0('<p>Hey there ', host(req), '!</p>')
    res[['Content-Type']] <- 'text/html'
    
    res
  })
}
\seealso{
\code{\link{body}}, \code{\link{host}}, \code{\link{type}},
  \code{\link{url}}
}

