% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response-coercion.R
\name{is.response}
\alias{is.response}
\title{Test if response}
\usage{
is.response(obj)
}
\arguments{
\item{obj}{An \R object}
}
\description{
\code{TRUE} if object is a response.
}
\examples{
# FALSE
is.response(logical(1))

# TRUE
is.response(response())

# FALSE
is.response(3030)
}

