% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route-class.R
\name{is.route}
\alias{is.route}
\title{Test if route}
\usage{
is.route(obj)
}
\arguments{
\item{obj}{Object to test}
}
\value{
\code{TRUE} if \code{obj} is of class \code{route}
}
\description{
Checks if \code{obj} is of class \code{route}.
}
\examples{
# One of these is not like the others
is.route('way')
is.route(90)
is.route(route('get', '^$', function(req, res) res$body('TRUE!')$send()))
}

