% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/callback-utils.R
\name{load_helpers}
\alias{load_helpers}
\title{Load callback helper functions}
\usage{
load_helpers(callback)
}
\description{
Loads a callback function's binding environment with a list of helpful
functions. These functions let uses access request information and set
response information more easily.
}
\keyword{internal}

