% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockup.R
\name{mockup}
\alias{mockup}
\title{Mockup a Route}
\usage{
mockup(r)
}
\arguments{
\item{r}{A \code{route} object.}
}
\value{
A route mockup will return the response object returned by the route handler.
However, if the method passed to the mockup is not handled by the route or if
the path is not handled by the route a 404 response object with a description
as the body is returned.
}
\description{
Creates a mockup of a route object. A mockup simulates what sort of response 
is generated by a route given a method, uri, and headers (optional).
}
\details{
Unfortunately, \code{httpuv} requires a least one header is specified. 
Therefore, \code{headers} defaults to and must have at least one field 
specified.
}
\examples{
logger <- route(
  'GET',
  '^',
  function(req) {
    print(req)
    response()
  }
)

logger_m <- mockup(logger)
logger_m('GET', '/yellow/brick/path')
logger_m('GET', '/phonday', headers = list(Accepts = 'text/html'))
}

