% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prairie.R
\docType{package}
\name{prairie}
\alias{prairie}
\alias{prairie-package}
\title{Grow Web Applications In R}
\description{
A web framework for R, prairie is based on existing
frameworks like Express (JS), Django (python), Wookie (lisp), and
Sinatra (ruby) and provides R users with a set of tools to create
web applications from scratch. Alternatively, users can coerce and
combine existing R objects into routes and applications.
}
\seealso{
\code{\link{app}}, \code{\link{route}}, \code{\link{request}}, and
\code{\link{response}}
}

