% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application.R
\name{print.application}
\alias{print.application}
\alias{format.application}
\title{Print an Application}
\usage{
\method{print}{application}(x, ...)

\method{format}{application}(x, ...)
}
\arguments{
\item{x}{An application.}

\item{\ldots}{Ignored.}
}
\description{
Prints an application.
}
\keyword{internal}
