% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{print.request}
\alias{print.request}
\title{Printing Requests}
\usage{
\method{print}{request}(x, ...)
}
\arguments{
\item{x}{A \code{request} object.}

\item{\ldots}{Ignored.}
}
\description{
Print a request.
}
\details{
Formats the request as an HTTP request.
}
\seealso{
\code{\link{request}}
}
\keyword{internal}

