% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{print.response}
\alias{print.response}
\alias{format.response}
\title{Printing Responses}
\usage{
\method{print}{response}(x, ...)

\method{format}{response}(x, ...)
}
\arguments{
\item{x}{Object of class \code{response}.}

\item{\ldots}{Ignored.}
}
\description{
Print a response object.
}
\details{
Formats the response as an HTTP response.
}
\seealso{
\code{\link{response}}
}
\keyword{internal}
