% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request-request-line.R
\name{query}
\alias{query}
\title{Request Query}
\usage{
query(x)
}
\arguments{
\item{x}{A request object.}
}
\description{
Get a request query.
}
\examples{
# This route prints out all query key, value pairs
route(
  'ALL',
  '^$',
  function(req) {
    if (!is.null(query(req))) {
      for (n in names(query(req))) {
        print(paste(n, query(req)[[n]]))
      }
    } else {
      print('The request did not contain a query')
    }

    response()
  }
)

}
\seealso{
Other HTTP.request.request.line: \code{\link{method}},
  \code{\link{uri}}
}

