% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{reason_phrase}
\alias{reason_phrase}
\title{Status Code Reason Phrase}
\usage{
reason_phrase(code)
}
\arguments{
\item{code}{An HTTP status code.}
}
\value{
The corresponding description of \code{code}, otherwise the empty
string.
}
\description{
Get the corresponding reason phrase for a status code.
}
\examples{
reason_phrase(200)
reason_phrase('404')

reason_phrase(531)

}
\keyword{internal}
