% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/callback-utils.R
\name{request-body}
\alias{body.request}
\alias{request-body}
\title{Retrieve the body of a request object}
\usage{
\method{body}{request}(.req)
}
\arguments{
\item{.req}{\code{response} object}
}
\description{
Retrieve the body of a request object
}
\examples{
library(magrittr)

dull() \%>\%
  post('^upload/csv/$', function(req, res) {
    file_name <- tempfile(pattern = '.csv', tempdir = 'public')

    # extract body from request and write to temp file
    writeBin(body(req), file_name)

    res \%>\%
      send('Upload successful!')
  })
}

