% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request-header.R
\name{request-headers}
\alias{[.request}
\alias{[[.request}
\alias{[[<-.request}
\alias{request-headers}
\title{HTTP Request Header Fields}
\usage{
\method{[}{request}(x, field)

\method{[[}{request}(x, field)

\method{[[}{request}(x, field) <- value
}
\arguments{
\item{x}{A \code{request} object.}

\item{field}{An HTTP request header field name.}

\item{value}{Value to assign to \code{field}.}
}
\description{
Get and header field values for \code{request} objects.
}
\details{
For more information regarding specific HTTP request header fields refer to 
\url{http://www.w3.org/Protocols/rfc2616/rfc2616-sec5.html}.
}
\examples{
req <- request()

req[['Accept']] <- 'text/*'
req[['From']] <- '127.0.0.1'
}

