% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request-line.R
\name{request-line}
\alias{method}
\alias{protocol}
\alias{request-line}
\alias{url}
\title{HTTP Request Line}
\usage{
method(x)

url(x)

protocol(x)
}
\arguments{
\item{x}{A \code{request} object.}
}
\description{
Get information about a request such as method type, url of the requested
resource, or the requset HTTP protocol.
}
\examples{
# A simple route to catch and log all incoming requests
route(
 'ALL',
 '^',
 function() {
   req <- request()
   res <- response()
   
   print(paste('[ LOG ]', Sys.time(), method(req), url(req), protocol(req)))
   
   body(res) <- 'Thanks for checking in'
   
   res
 }
)
}

