% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/request.R
\docType{class}
\name{request}
\alias{request}
\title{Request class}
\format{An R6 class object.}
\usage{
request
}
\description{
The request class holds meta information about a request received for a
particular application resource (URI). This information is pulled from an
underlying Rook environment. A request object is passed to the corresponding
callback function when a URI is requested.
}
\section{Methods}{

\itemize{
 \item \code{get_header_field(field)}:
   Retrieve a HTTP header field value from the underlying Rook environment
 \item \code{has_content_type(type)}:
   TRUE if the request content type is \code{type}
 \item \code{get_method()}:
   Return request HTTP method
 \item \code{get_body()}:
   Return the request body
 \item \code{get_ip()}:
   Return the remote ip address of the request
 \item \code{get_port()}:
   Return the remove host address of the request
 \item \code{get_host_name()}:
   Return the HTTP remote host
 \item \code{get_params()}:
   Return captured URI parameters as a list, list names are capture group name
}
}
\keyword{internal}

