% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/response.R
\docType{data}
\name{response}
\alias{response}
\title{}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x10f83ffc0> 
 - attr(*, "name")= chr "response_generator"
}}
\usage{
response
}
\description{

}
\keyword{datasets}

