% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route-class.R
\docType{class}
\name{route-class}
\alias{route-class}
\alias{route__}
\title{Route class}
\format{An R6 class object.}
\usage{
route__
}
\description{
This class is a glorified container for a path and handler. Some handy
utility functions are included as private methods. As dull middleware begins
to take shape this class will evolve more.
}
\keyword{internal}

