% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application.R
\name{run}
\alias{run}
\title{Start Up a Prairie Application}
\usage{
run(app, host, port)
}
\arguments{
\item{app}{An object with class \code{application}.}

\item{host}{A character string specifying the host name.}

\item{port}{An integer specifying the port number.}
}
\description{
Run an application at \code{host} on \code{port}.
}

