% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simpleHtmlTable}
\alias{simpleHtmlTable}
\title{Generate a minimal HTML table from a data frame
 
 This internal function is used by as.response()
 to encode a data frame into its HTML representation.
 The encoding is intentionally very minimal.
 Richer encoding would require either a complicated
 function or importing a heavy-weight package.}
\usage{
simpleHtmlTable(dfrm, class = NULL, id = NULL)
}
\arguments{
\item{dfrm}{a data frame}

\item{class}{an optional CSS class for the table}

\item{id}{an optional CSS ID for the table}
}
\value{
an HTML representation of a table (character)
}
\description{
Generate a minimal HTML table from a data frame
 
 This internal function is used by as.response()
 to encode a data frame into its HTML representation.
 The encoding is intentionally very minimal.
 Richer encoding would require either a complicated
 function or importing a heavy-weight package.
}
