% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application.R
\name{start}
\alias{start}
\title{Run Prairie Application}
\usage{
start(app, host, port)
}
\arguments{
\item{app}{An \code{application} object.}

\item{host}{A character vector specifying the host name.}

\item{port}{A numeric specifying the port number.}
}
\description{
Starts an \code{\link{app}} at \code{host} on \code{port}.
}

