% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{status}
\alias{status}
\alias{status.response}
\alias{status<-}
\alias{status<-.response}
\title{HTTP Response Status Code}
\usage{
status(x)

status(x) <- value

\method{status}{response}(x)

\method{status}{response}(x) <- value
}
\arguments{
\item{x}{An \R object}

\item{value}{HTTP status code, 1xx through 5xx}
}
\description{
Get or set the status code of a \code{response} object.
}

