% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tags.R
\name{templating}
\alias{a}
\alias{abbr}
\alias{acronym}
\alias{address}
\alias{applet}
\alias{area}
\alias{article}
\alias{aside}
\alias{audio}
\alias{b}
\alias{base}
\alias{basefont}
\alias{bdi}
\alias{bdo}
\alias{big}
\alias{blockquote}
\alias{body.character}
\alias{br}
\alias{button}
\alias{canvas}
\alias{caption}
\alias{center}
\alias{cite}
\alias{code}
\alias{col}
\alias{colgroup}
\alias{datalist}
\alias{dd}
\alias{del}
\alias{details}
\alias{dfn}
\alias{dialog}
\alias{dir}
\alias{div}
\alias{dl}
\alias{doctype}
\alias{dt}
\alias{em}
\alias{embed}
\alias{fieldset}
\alias{figcaption}
\alias{figure}
\alias{font}
\alias{footer}
\alias{form}
\alias{frame}
\alias{frameset}
\alias{h1}
\alias{h2}
\alias{h3}
\alias{h4}
\alias{h5}
\alias{h6}
\alias{head}
\alias{header}
\alias{hr}
\alias{html}
\alias{i}
\alias{iframe}
\alias{img}
\alias{input}
\alias{ins}
\alias{kbd}
\alias{keygen}
\alias{label}
\alias{legend}
\alias{li}
\alias{link}
\alias{main}
\alias{map}
\alias{mark}
\alias{menu}
\alias{menuitem}
\alias{meta}
\alias{meter}
\alias{nav}
\alias{noframes}
\alias{noscript}
\alias{object}
\alias{ol}
\alias{optgroup}
\alias{option}
\alias{output}
\alias{p}
\alias{param}
\alias{pre}
\alias{progress}
\alias{q}
\alias{rp}
\alias{rt}
\alias{ruby}
\alias{s}
\alias{samp}
\alias{script}
\alias{section}
\alias{select}
\alias{small}
\alias{source}
\alias{span}
\alias{strike}
\alias{strong}
\alias{style}
\alias{sub}
\alias{summary}
\alias{sup}
\alias{table}
\alias{tag}
\alias{tbody}
\alias{td}
\alias{templating}
\alias{textarea}
\alias{tfoot}
\alias{th}
\alias{thead}
\alias{time}
\alias{title}
\alias{tr}
\alias{track}
\alias{tt}
\alias{u}
\alias{ul}
\alias{var}
\alias{video}
\alias{wbr}
\title{Create an HTML tag string}
\usage{
tag(name, args)

doctype(...)

a(...)

abbr(...)

acronym(...)

address(...)

applet(...)

area(...)

article(...)

aside(...)

audio(...)

b(...)

base(...)

basefont(...)

bdi(...)

bdo(...)

big(...)

blockquote(...)

\method{body}{character}(...)

br(...)

button(...)

canvas(...)

caption(...)

center(...)

cite(...)

code(...)

col(...)

colgroup(...)

datalist(...)

dd(...)

del(...)

details(...)

dfn(...)

dialog(...)

dir(...)

div(...)

dl(...)

dt(...)

em(...)

embed(...)

fieldset(...)

figcaption(...)

figure(...)

font(...)

footer(...)

form(...)

frame(...)

frameset(...)

h1(...)

h2(...)

h3(...)

h4(...)

h5(...)

h6(...)

head(...)

header(...)

hr(...)

html(...)

i(...)

iframe(...)

img(...)

input(...)

ins(...)

kbd(...)

keygen(...)

label(...)

legend(...)

li(...)

link(...)

main(...)

map(...)

mark(...)

menu(...)

menuitem(...)

meta(...)

meter(...)

nav(...)

noframes(...)

noscript(...)

object(...)

ol(...)

optgroup(...)

option(...)

output(...)

p(...)

param(...)

pre(...)

progress(...)

q(...)

rp(...)

rt(...)

ruby(...)

s(...)

samp(...)

script(...)

section(...)

select(...)

small(...)

source(...)

span(...)

strike(...)

strong(...)

style(...)

sub(...)

summary(...)

sup(...)

table(...)

tbody(...)

td(...)

textarea(...)

tfoot(...)

th(...)

thead(...)

time(...)

title(...)

tr(...)

track(...)

tt(...)

u(...)

ul(...)

var(...)

video(...)

wbr(...)
}
\arguments{
\item{name}{The tag name}

\item{args}{Additional HTML attributes and tag values}

\item{...}{HTML attributes and values for the specific tag}
}
\description{
A handy function to create HTML tags. Named arguments are interpreted as HTML
tag attributes and unnamed arguments are used as the value. Mulitple unnamed
values are concatenated together as a single value in the order they appear.
}
\details{
Most of the time one will want to use the default wrapper functions like
\code{a}, \code{body} or \code{div}, see below for the complete list.
However, \code{tag} allows for one to work with custom tags. Because of this,
\code{tag} naiveley assumes the named arguments passed as \code{...} are
appropriate attributes for tag \code{name}.
}
\examples{
## Basic tags
h3("Third Time's a Charm!")
b("Emphasize this and that")

## Nesting tags
body(
  h1("I'm Number One!"),
  p("A biased analysis of life as an eldest child.")
)

## Custom tags
tag("hmmm", type = vague, "Was this a good idea?")
}

