% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request-header.R
\name{type}
\alias{type}
\title{HTTP Request Content Type}
\usage{
type(req)
}
\arguments{
\item{req}{A \code{request} object}
}
\description{
Get the content types of a request.
}
\seealso{
\code{\link{body}}, \code{\link{host}}, \code{\link{type}},
  \code{\link{url}}
}

