% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request-line.R
\name{url}
\alias{url}
\title{HTTP Request URL}
\usage{
url(req)
}
\arguments{
\item{req}{A \code{request} object}
}
\description{
Get the URL for a request.
}
\seealso{
\code{\link{body}}, \code{\link{host}}, \code{\link{type}},
  \code{\link{url}}
}

