% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application.R
\name{app}
\alias{app}
\title{Create application}
\usage{
app(...)
}
\arguments{
\item{\ldots}{Route objects}

\item{host}{host address to run the application on}

\item{port}{port to run the application on}
}
\description{
Applications in prairie are naively defined as a \emph{list of 
\link{routes}}. The application handles dispatching HTTP requests to the 
routes and responding with HTTP responses. However, response building happens
within each route.
}

