% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application.R
\docType{class}
\name{application}
\alias{application}
\title{Application class object}
\format{An R6 class object.}
\usage{
application
}
\description{
An R6 class which is the backbone of the dull package.
}
\keyword{internal}

