% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{as.request}
\alias{as.request}
\alias{as.request.environment}
\title{Coerce Rook Environments to Requests}
\usage{
as.request(x)

\method{as.request}{environment}(x)
}
\arguments{
\item{x}{An \R object.}
}
\description{
Internally, this function is used to coerce the request environment objects
\code{httpuv} passes to an application's \code{call} function. Request
environment objects are coerced to objects.
}
\examples{
e <- new.env(parent = emptyenv())

e$REQUEST_METHOD <- 'GET'
e$PATH_INFO <- '/file/download'
e$HTTP_ACCEPT <- 'application/json'
e$HTTP_CONTENT_LENGTH <- '0'

req <- as.request(e)
is.request(req)  # TRUE

method(req)
uri(req)
req[['Accept']]
req[['Content-Length']]

}
\seealso{
\code{\link{request}}
}
\keyword{internal}
