% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{as.response}
\alias{as.response}
\alias{is.response}
\title{Coerce Objects to Response}
\usage{
as.response(x)

is.response(x)
}
\arguments{
\item{x}{Any \R object.}
}
\description{
TBD, see jsonify.
}
\examples{
is.response(logical(1))
is.response(response())
is.response(3030)
}

