% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{as.response}
\alias{as.response}
\alias{as.response.character}
\alias{as.response.data.frame}
\alias{as.response.matrix}
\alias{is.response}
\title{Coerce Objects to Responses}
\usage{
as.response(x, ...)

\method{as.response}{character}(x, directory = "views", collapse = "\\n",
  ...)

\method{as.response}{data.frame}(x, format = "json", ...)

\method{as.response}{matrix}(x, ...)

is.response(x)
}
\arguments{
\item{x}{Any \R object.}

\item{directory}{A character string specifying the system folder of the file
\code{x}.}

\item{collapse}{A character string specifying how to collapse the lines read
from \code{x}.}

\item{\ldots}{Additional arguments passed on to methods.}
}
\description{
The \code{as.response} functions makes serving up R objects as server
responses straightfoward.
}
\details{
\code{as.response.character} expects \code{x} is a character string
specifying a file name. The default directory for the file is "views", but a
different path may be specified with the \code{directory} argument. If the
file exists the contents are read and set as the response body. The response
Content-Type is guessed from the file extension using
\code{\link[mime]{guess_type}}.

\code{as.response.data.frame} coerces and serves up a data frame as JSON. The
data frame is coerced using the \code{\link{as.json}} function and additional
arguments may be passed to \code{as.json} using \code{\ldots}.
}
\examples{

is.response(logical(1))
is.response(response())
is.response(3030)
}
