% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.R
\name{as.route}
\alias{as.route}
\alias{as.route.route}
\alias{as.route.character}
\alias{as.route.list}
\title{Coercing Objects to Routes}
\usage{
as.route(x, ...)

\method{as.route}{route}(x, ...)

\method{as.route}{character}(x, directory = "routes", ...)

\method{as.route}{list}(x, ...)
}
\arguments{
\item{x}{An \R object.}

\item{directory}{System path to the folder containing the route file.}

\item{\ldots}{Additional arguments passed on to methods.}
}
\value{
A \code{route} object.
}
\description{
The function \code{as.route} provides an alternative means of creating an
application route.
}
\details{
If \code{x} is a list, \code{x} must have the following named elements:
\code{method}, \code{path}, and \code{handler}.

If \code{x} is a character vector, \code{x} is interpreted as a file name.
The file must contain a route defined using the \code{route} function. The
default directory for route files is "routes", but a different folder may be
specified with the \code{directory} argument.

The S3 generic function \code{as.route} is exported by prairie to encourage
creation of \code{as.route.*} functions. Custom \code{as.route} functions
allow users to coerce their classes to routes and quickly serve them over
HTTP.
}
\examples{
# Easily reuse routes and keep applications
# modular by storing routes in separate files.

tmp <- tempfile()
writeLines(
  'route("GET", "^$", function(req) response())\\n',
  con = tmp
)

as.route(tmp, dir = '')

file.remove(tmp)

# as.route.list is a minimal wrapper
# around route()

route(
  'POST',
  '^$',
  function(req) {
    response()
  }
)

as.route(
  list(
    method = 'POST',
    path = '^$',
    handler = function(req) {
      response()
    }
  )
)
}
\seealso{
\code{\link{route}}
}
