% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{body}
\alias{body}
\alias{body.request}
\alias{body.response}
\alias{body<-}
\alias{body<-.response}
\title{HTTP Message Body}
\usage{
body(x)

body(x) <- value

\method{body}{request}(x)

\method{body}{response}(x)

\method{body}{response}(x) <- value
}
\arguments{
\item{x}{An \R object.}

\item{value}{An \R object, complex objects are converted to JSON, see details.}
}
\value{
A character string if the request contains a message body, otherwise
\code{NULL}.
}
\description{
Get the message body of a \code{request} or \code{response} object. The body
of \code{response} objects may be set.
}
\examples{
route(
  'post',
  '^transmogrify/json$',
  function() {
    req <- request()
    res <- response()

    # get body of request
    raw_body <- body(req)
    data <- jsonlite::fromJSON(raw_body)

    mogrified <- lapply(
      data,
      function(col) paste0(as.character(col), '!!!')
    )

    status(res) <- 200

    # set body of response as data.frame
    # coerced to JSON string with as.JSON
    body(res) <- mogrified

    res
  }
)
}
\seealso{
Additional \code{request} functions: \code{\link{body}}, \code{\link{host}},
\code{\link{type}}, \code{\link{url}}

For more information about HTTP requests please refer to the
\href{http://www.w3.org/Protocols/rfc2616/rfc2616-sec5.html}{Request}
section of www.w3.org.
}

