% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{body}
\alias{body}
\alias{body<-}
\title{HTTP Message Body}
\usage{
body(x)

body(x) <- value
}
\arguments{
\item{x}{An \R object.}

\item{value}{The response body, if JSON the \code{Content-Type} is set to
\code{application/json}.}
}
\value{
\code{body} resturns a character string if the request or response contains a
message body, otherwise \code{NULL}.

\code{body<-} invisibly returns the response.
}
\description{
Get the message body of a request or response or set the body of a response.
}
\details{
For more information about HTTP requests please refer to the
\href{http://www.w3.org/Protocols/rfc2616/rfc2616-sec5.html}{Request} section
of www.w3.org.
}
\examples{
transmog <- route(
  'POST',
  '^transmogrify/json$',
  function(req) {
    res <- response()

    # get body of request
    raw_body <- body(req)
    data <- jsonlite::fromJSON(raw_body)

    mogrified <- lapply(
      data,
      function(col) paste0(as.character(col), '!!!')
    )

    status(res) <- 200
    body(res) <- as.json(mogrified)

    res
  }
)
}
