% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dull.R
\docType{package}
\name{dull-package}
\alias{dull-package}
\title{dull, building web applications too easily}
\description{
\code{dull} provides a straight-forward syntax to build web applications in 
R. Functions like \link{get}, \link{post}, and \link{put} add route callbacks
to a \link{dull} application.
}
\details{
For information regarding route callbacks and route URI specification refer
to \link{dull}.
}

