% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dull.R
\docType{package}
\name{dull-package}
\alias{dull}
\alias{dull-package}
\title{A Web Framework For R}
\description{
A straight forward web framework for creating modular web
applications with R. Based on existing frameworks like Express (JS),
Wookie (lisp), and Sinatra (ruby), dull provides R users with a set
of tools to create powerful, simple-to-build web applications.
}
\seealso{
\code{\link{app}}, \code{\link{route}}, \code{\link{request}}, and
\code{\link{response}}
}

