% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request-request-line.R
\name{method}
\alias{method}
\title{Request Method}
\usage{
method(x)
}
\arguments{
\item{x}{A \code{request} object.}
}
\description{
Get information about a request such as method type, the requested resource
uri, the query component, or the requset HTTP protocol.
}
\examples{

methodical <- route(
 c('GET', 'POST'),
 '^',
 function(req) {
   res <- response()
   
   if (method(req) == 'GET') {
     # handle when method is GET
     body(res) <- 'I got ya'  
     
   } else {
     body(res) <- 'Washingtong or Huffington?'
     
   }
   
   res
 }
)

methodical_m <- mockup(methodical)

res <- methodical_m('GET', '/')
res

res <- methodical_m('POST', '/')
body(res)
}
\seealso{
Other HTTP.request.request.line: \code{\link{query}},
  \code{\link{uri}}
}

