% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockup.R
\name{mockup}
\alias{mockup}
\alias{mockup.route}
\alias{print.mockup}
\title{Mockup a Route}
\usage{
mockup(x)

\method{mockup}{route}(x)

\method{print}{mockup}(x, ...)
}
\arguments{
\item{x}{Any \R object}
}
\description{
Creates a mockup of a route object. A mockup simulates a route's response to
a particular method and resource combination.
}
\examples{
logger <- route(
  'GET',
  '^',
  function(req) {
    print(req)
    
    response()
  }
)

logger_m <- mockup(logger)
logger_m('GET', '/yellow/brick/path')
logger_m('GET', '/pho', headers = list(Accepts = 'text/html'))
}

