% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockup.R
\name{mockup}
\alias{mockup}
\title{Mockup a Route}
\usage{
mockup(r)
}
\arguments{
\item{r}{A \code{route} object.}
}
\description{
Creates a mockup of a route object. A mockup simulates a route's response to
a particular method and resource combination.
}
\examples{
logger <- route(
  'GET',
  '^',
  function(req) {
    print(req)
    response()
  }
)

logger_m <- mockup(logger)
logger_m('GET', '/yellow/brick/path')
logger_m('GET', '/phonday', headers = list(Accepts = 'text/html'))
}

