% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{print.request}
\alias{format.request}
\alias{print.request}
\title{Printing Requests}
\usage{
\method{print}{request}(x, ...)

\method{format}{request}(x, ...)
}
\arguments{
\item{x}{A request object.}

\item{\ldots}{Ignored.}
}
\description{
Print a request object.
}
\examples{
print(request())

}
\seealso{
\code{\link{request}}
}
\keyword{internal}

