% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.R
\name{print.route}
\alias{format.route}
\alias{print.route}
\title{Printing Routes and Mockups}
\usage{
\method{print}{route}(x, ...)

\method{format}{route}(x, ...)
}
\arguments{
\item{x}{A route or route mockup.}

\item{\ldots}{Ignored.}
}
\description{
Prints a route or mockup. Printing a mockup prints the underlying route.
}
\examples{
route(
  c('GET', 'POST'),
  '^path$',
  function(req) {
    response()
  }
)

route(
  'put',
  '^another/path$',
  function(req) {
    response()
  }
)
}
\seealso{
\code{\link{route}}, \code{\link{mockup}}
}
\keyword{internal}

