% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.R, R/mockup.R
\name{print.route}
\alias{print.mockup}
\alias{print.route}
\title{Printing Routes and Mockups}
\usage{
\method{print}{route}(x, ...)

\method{print}{mockup}(x, ...)
}
\arguments{
\item{x}{A \code{route} or route \code{mockup}.}
}
\description{
Prints a route or mockup. A mockup prints the underlying route object.
}
\seealso{
\code{\link{route}}, \code{\link{mockup}}
}

