% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{reason_phrase}
\alias{reason_phrase}
\title{Status Code Reason Phrase}
\usage{
reason_phrase(status)
}
\arguments{
\item{status}{An HTTP status code.}
}
\value{
If \code{status} is not found, then the empty string is returned.
}
\description{
Get the corresponding reason phrase for a particular status code.
}
\examples{
reason_phrase(200)
reason_phrase('404')

reason_phrase(531)
}
\keyword{internal}

