% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request-header.R
\name{request-headers}
\alias{[.request}
\alias{[[.request}
\alias{request-headers}
\title{Request Header Fields}
\usage{
\method{[[}{request}(x, field)

\method{[}{request}(x, field)
}
\arguments{
\item{x}{A \code{request} object.}

\item{field}{An HTTP request header field name.}
}
\description{
To get the values of request header fields [ or [[ may be used to get a 
single or multiple values respectively. Request field names are 
case-insensitive, so \code{"Accept"} and \code{"accept"} are equivalent.
}
\details{
For more information regarding specific HTTP request header fields please
refer to \url{https://tools.ietf.org/html/rfc2616#section-5.3}.
}
\examples{
req <- request()

req[['Accept']] # NULL
req[['From']]   # NULL

route(
  'POST',
  '^$',
  function(req) {
    # Now there will be values
    print(req[['Accept']])
    print(req[['From']])
    
    response()
  }
)
}

