% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request-header.R
\name{request-headers}
\alias{[.request}
\alias{[[.request}
\alias{request-headers}
\title{Request Header Fields}
\usage{
\method{[[}{request}(x, field)

\method{[}{request}(x, field)
}
\arguments{
\item{x}{A \code{request} object.}

\item{field}{An HTTP request header field name.}
}
\description{
To get the values of request header fields [ or [[ may be used to get a 
single or multiple values respectively. Request field names are 
case-sensitive.
}
\details{
For more information regarding HTTP request header fields please refer to
\url{https://tools.ietf.org/html/rfc2616#section-5.3}.
}
\examples{
req <- request()

req[['Accept']] # NULL
req[['From']]   # NULL, boring

checkin <- route(
  'POST',
  '^$',
  function(req) {
    print(req[['Accept']])
    print(req[['From']])
    
    response()
  }
)

checkin_m <- mockup(checkin)

# More interesting output
checkin_m(
  'POST', 
  '/', 
  headers = list(
    Accept = 'text/html', 
    From = 'Russia w/ Love'
  )
)
}

