% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{request}
\alias{request}
\title{Create an HTTP Request}
\usage{
request()
}
\description{
Request objects store relevant information sent from the client to the server
as a part of an HTTP request. Request objects are not typically explicitly 
created. Instead, a request object is passed as an argument to a route
handler.
}
\details{
Request objects contain the following information.

\subsection{method:}{
  
}
}

