% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response-headers.R
\name{response-headers}
\alias{[.response}
\alias{[[.response}
\alias{[[<-.response}
\alias{response-headers}
\title{HTTP Response Headers}
\usage{
\method{[}{response}(res, name)

\method{[[}{response}(res, name)

\method{[[}{response}(res, name) <- value
}
\arguments{
\item{res}{A \code{response} object}

\item{name}{HTTP response header field(s), case sensitive}

\item{value}{Value to assign to field \code{name}}
}
\value{
A list of corresponding header field values. If \code{name} is not set for 
\code{res} then \code{NULL} is returned.
}
\description{
Within prairie, getting and setting the fields of the \code{\link{response}} 
header is much like assigning values to lists. One can use \code{[}, 
\code{[[} to extract one or more fields or replace a single 
field.
}
\examples{
# create new response
res <- response()

res[["Content-Type"]] <- "text/html"
}
\seealso{
For more information regarding specific HTTP response header fields refer to 
\url{http://www.w3.org/Protocols/rfc2616/rfc2616-sec6.html}.
}

