% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response-headers.R
\name{response-headers}
\alias{$.response}
\alias{$<-.response}
\alias{[.response}
\alias{[[.response}
\alias{[[<-.response}
\alias{response-headers}
\title{HTTP Response Headers}
\usage{
\method{[}{response}(res, name)

\method{[[}{response}(res, name)

\method{[[}{response}(res, name) <- value

\method{$}{response}(res, name)

\method{$}{response}(res, name) <- value
}
\arguments{
\item{res}{A \code{response} object}

\item{name}{HTTP response header field(s), case sensitive}

\item{value}{Value to assign to field \code{name}}
}
\description{
Within prairie, getting and setting the fields of the \code{\link{response}} 
header is much like assigning values to lists. One can use \code{[}, 
\code{[[}, and \code{$} to extract or replace a single field or multiple 
fields.
}
\details{
Use \code{[[} to access fields with non-alphanumeric characters, such as 
\emph{Content-Type} which contains \emph{-}. If your heart is set on 
accessing fields with \code{$} backticks may also be used.
}
\seealso{
For more information regarding specific HTTP response header fields refer to
\url{http://www.w3.org/Protocols/rfc2616/rfc2616-sec6.html}.
}

