% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{response}
\alias{response}
\title{Create an HTTP response}
\usage{
response()
}
\description{
A response object represents the HTTP response returned by a route handler.
}
\examples{
# simple 301 response
res_301 <- resposne()
    
res[['Connection']] <- 'close'
status(res_301) <- 301
body(res_301) <- 'Sorry, this page has moved'

# simple 404 response
res_404 <- response()

res[['Connection']] <- 'close'
status(res_404) <- 404
body(res_404) <- 'Sorry, page not found'
}
\seealso{
\code{\link{route}}, \code{\link{req}}
}

