% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/route.R
\docType{class}
\name{route-class}
\alias{route}
\alias{route-class}
\title{Simple route class}
\format{An R6 class object.}
\usage{
route
}
\description{
This class is a glorified container for a url and the corresponding callback
function. Some handy utility functions are included as private methods. As
dull middleware begins to take shape this class will evolve more.
}
\section{Methods}{

\itemize{
 \item \code{uri_matches(path)}: TRUE if the route
  uri, as a regular expression, matches path
 \item \code{assign_callback(method, callback)}: set the callback function for a
  specific HTTP method
 \item \code{callback_for(method)}: return the callback
  function for a specific HTTP method
}
}
\keyword{internal}

