% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sharpen.R
\name{sharpen}
\alias{expect_get}
\alias{expect_post}
\alias{expect_put}
\alias{expect_response}
\alias{sharpen}
\title{Test a dull application}
\usage{
expect_response(.app, method, uri, status, ...)

expect_get(.app, uri, status, ...)

expect_post(.app, uri, status, ...)

expect_put(.app, uri, status, ...)
}
\arguments{
\item{.app}{A dull application}

\item{method}{A character vector of the HTTP method with which to query
\code{url}}

\item{uri}{A character vector of the resource path to query}

\item{status}{An integer, the expected status}

\item{\dots}{Additional parameters to pass to \code{\link[httr]{VERB}}}
}
\description{
A set of functions to expand upon the \code{expect_*} functions of the
\code{testthat} package.
}
\details{
These functions are currently very limited and would be considered
experimental if not for the \code{testthat}. There is plenty more to come in the
future.

\code{uri} must begin with a \dQuote{/} and is used to construct the complete
URL queried. The host name and port used are extracted from the \code{.app}
object.
}

