% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application.R
\name{start}
\alias{start}
\title{Run a Prairie Application}
\usage{
start(app, host, port)
}
\arguments{
\item{app}{An object with class \code{application}.}

\item{host}{A character string specifying the host name.}

\item{port}{An integer specifying the port number.}
}
\description{
Start an application running at \code{host} on \code{port}.
}

