% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{status}
\alias{status}
\alias{status<-}
\title{HTTP Response Status Code}
\usage{
status(x)

status(x) <- value
}
\arguments{
\item{x}{A response object.}

\item{value}{An HTTP status code, \code{1xx} through \code{5xx}, see
\code{\link{response}} for details.}
}
\description{
Get or set the status code of a response object.
}
\examples{
# create a new response
res <- response()

# default response status
status(res)  # 200

# set the response status
status(res) <- 301
status(res)

}
